requirements.txt
matplotlib
numpy
pandas
geopandas
libpysal
seaborn
contextily
pointpats
sklearn
descartes